raw SQL should be a last resort
QuerySet.extra() should be used as a last resort, it is no longer maintained and the Django maintainers intend on removing it in the future.
https://docs.djangoproject.com/en/1.11/ref/models/querysets/#extra

The custom SQL has been replaced with ConvertFrom() and annotations. This also removes the requirement for a database Specific SQLconnector() as the database specific code is handled in the ConvertFrom() function class.